error handling implementation

+++ src/services/university.service.js
@@ -12,13 +12,29 @@ const postUniversity = async (university) => {
   return newUniversity;
 };
 
+const checkIfExists = async (id) => {
+  const exist = await getUniversityById(id);
+
+  if (!exist) return false;
+
+  return true;
+};
+
 const putUniversity = async (id, data) => {
+  const exist = await checkIfExists(id);
+
+  if (!exist) return false;
+
   const updatedTask = await UniversityModel.findByIdAndUpdate(id, data);
 
   return { ...updatedTask._doc, ...data };
 };
 
 const destroyUniversity = async (id) => {
+  const exist = await checkIfExists(id);
+
+  if (!exist) return false;
+
   const deletedUniversity = await UniversityModel.findByIdAndDelete(id);
 
   return deletedUniversity;
